AC3 Did a round of code refactoring and fill up some of the sections with basic information
