+++ src/router/index.js
@@ -1,7 +1,7 @@
 import Vue from 'vue'
 import VueRouter from 'vue-router'
 import homeView from '../views/home.vue'
-import paletteView from '../views/configuration/palette.vue'
+import paletteView from '../views/configuration/palette-configuration.vue'
 
 Vue.use(VueRouter)
 
@@ -17,11 +17,11 @@ const routes = [
     // route level code-splitting
     // this generates a separate chunk (about.[hash].js) for this route
     // which is lazy-loaded when the route is visited.
-    component: () => import(/* webpackChunkName: "about" */ '../views/About.vue')
+    component: () => import(/* webpackChunkName: "about" */ '../views/about.vue')
   },
   {
-    path: '/configuration/palette',
-    name: 'configuration-palette',
+    path: '/configuration/palette-configuration',
+    name: 'palette-configuration',
     component: paletteView
   }
 ]
