+++ src/scss/App.scss
@@ -6,6 +6,8 @@
   -moz-osx-font-smoothing: grayscale;
   text-align: center;
   color: green;
+  min-height: calc(100vh - 2vh);
+  min-width: calc(100vw - 1vw);
 }
 
 #nav {
@@ -19,4 +21,10 @@
 
 #nav a.router-link-exact-active {
   color: #42b983;
+}
+
+*,
+::after,
+::before {
+  box-sizing: border-box;
 }
